Remove Old Dev UI: Core Configuration
