+++ extensions/vertx-http/deployment/src/main/java/io/quarkus/vertx/http/deployment/devmode/console/DevConsoleFailedStartHandler.java
@@ -2,11 +2,12 @@ package io.quarkus.vertx.http.deployment.devmode.console;
 
 import io.quarkus.dev.config.CurrentConfig;
 import io.quarkus.dev.spi.DeploymentFailedStartHandler;
+import io.quarkus.devui.deployment.menu.ConfigurationProcessor;
 
 public class DevConsoleFailedStartHandler implements DeploymentFailedStartHandler {
     @Override
     public void handleFailedInitialStart() {
         DevConsoleProcessor.initializeVirtual();
-        CurrentConfig.EDITOR = ConfigEditorProcessor::updateConfig;
+        CurrentConfig.EDITOR = ConfigurationProcessor::updateConfig;
     }
 }
