+++ extensions/vertx-http/deployment/src/main/resources/dev-templates/tags/actionableExtension.html
@@ -24,10 +24,5 @@
                 {/if}
             </p>
         </div>
-        {#if it.metadata.config??}
-        <div class="card-footer border-0 bg-white">
-            <a href='io.quarkus.quarkus-vertx-http/config?filterByExtension={it.name}&filterConfigKeys={it.metadata.config.toString().replace("[", "").replace("]", "").replace(" ", "")}' class='float-right text-secondary' title='Config Editor'><i class='fas fa-edit'></i></a>
-        </div>
-        {/if}
     </div>
 </div>
