+++ extensions/vertx-http/deployment/src/test/java/io/quarkus/vertx/http/devconsole/DevConsoleCorsTest.java
@@ -24,7 +24,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", origin)
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", methods)
@@ -39,7 +39,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", origin)
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", methods)
@@ -54,7 +54,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", origin)
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", methods)
@@ -69,7 +69,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", origin)
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", methods)
@@ -83,7 +83,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", "https://quarkus.io/http://localhost")
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(403)
                 .header("Access-Control-Allow-Origin", nullValue())
                 .header("Access-Control-Allow-Methods", nullValue())
@@ -97,7 +97,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", "http://localhost:8080/devui")
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(403)
                 .header("Access-Control-Allow-Origin", nullValue())
                 .body(emptyOrNullString());
@@ -110,7 +110,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", "http://127.0.0.1:8080/devui")
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(403)
                 .header("Access-Control-Allow-Origin", nullValue())
                 .body(emptyOrNullString());
@@ -123,7 +123,7 @@ public class DevConsoleCorsTest {
                 .header("Origin", "http://localhost")
                 .header("Access-Control-Request-Method", methods)
                 .when()
-                .options("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .options("q/dev-ui/configuration-form-editor").then()
                 .statusCode(403)
                 .header("Access-Control-Allow-Origin", nullValue())
                 .body(emptyOrNullString());
@@ -135,7 +135,7 @@ public class DevConsoleCorsTest {
         RestAssured.given()
                 .header("Origin", origin)
                 .when()
-                .get("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .get("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", nullValue())
@@ -148,7 +148,7 @@ public class DevConsoleCorsTest {
         RestAssured.given()
                 .header("Origin", origin)
                 .when()
-                .get("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .get("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", nullValue())
@@ -161,7 +161,7 @@ public class DevConsoleCorsTest {
         RestAssured.given()
                 .header("Origin", origin)
                 .when()
-                .get("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .get("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", nullValue())
@@ -174,7 +174,7 @@ public class DevConsoleCorsTest {
         RestAssured.given()
                 .header("Origin", origin)
                 .when()
-                .get("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .get("q/dev-ui/configuration-form-editor").then()
                 .statusCode(200)
                 .header("Access-Control-Allow-Origin", origin)
                 .header("Access-Control-Allow-Methods", nullValue())
@@ -186,7 +186,7 @@ public class DevConsoleCorsTest {
         RestAssured.given()
                 .header("Origin", "https://quarkus.io/http://localhost")
                 .when()
-                .get("q/dev-v1/io.quarkus.quarkus-vertx-http/config").then()
+                .get("q/dev-ui/configuration-form-editor").then()
                 .statusCode(403)
                 .header("Access-Control-Allow-Origin", nullValue())
                 .body(emptyOrNullString());
