+++ extensions/vertx-http/runtime/src/main/java/io/quarkus/devui/runtime/config/ConfigDevUIRecorder.java
@@ -14,12 +14,11 @@ import org.eclipse.microprofile.config.ConfigProvider;
 import org.eclipse.microprofile.config.spi.ConfigSource;
 
 import io.quarkus.runtime.annotations.Recorder;
-import io.quarkus.vertx.http.runtime.devmode.ConfigDescription;
 import io.smallrye.config.ConfigValue;
 import io.smallrye.config.SmallRyeConfig;
 
 @Recorder
-public class ConfigDevUiRecorder {
+public class ConfigDevUIRecorder {
     private static final String QUOTED_DOT = "\".\"";
     private static final String QUOTED_DOT_KEY = "$$QUOTED_DOT$$";
 
@@ -27,8 +26,8 @@ public class ConfigDevUiRecorder {
     private static Set<String> devServicesConfig;
 
     public void registerConfigs(final List<ConfigDescription> configDescriptions, final Set<String> devServicesConfig) {
-        ConfigDevUiRecorder.configDescriptions = configDescriptions;
-        ConfigDevUiRecorder.devServicesConfig = devServicesConfig;
+        ConfigDevUIRecorder.configDescriptions = configDescriptions;
+        ConfigDevUIRecorder.devServicesConfig = devServicesConfig;
     }
 
     public Supplier<ConfigDescriptionBean> configDescriptionBean() {
